Rename the program module to store
This is for consistency with similar nomenclature in all other crates.
