+++ cli/src/lib.rs
@@ -298,6 +298,6 @@ impl console::Console for TextConsole {
 pub fn run_repl_loop(dir: &Path) -> io::Result<i32> {
     let console = Rc::from(RefCell::from(TextConsole::from_stdio()?));
     let store = FileStore::new(dir);
-    let store = Rc::from(RefCell::from(endbasic_core::program::DemoStoreOverlay::new(store)));
+    let store = Rc::from(RefCell::from(endbasic_core::store::DemoStoreOverlay::new(store)));
     block_on(endbasic_core::repl::run_repl_loop(console, store))
 }
