+++ core/src/store.rs
